+++ OMGBlockz/trunk/src/com/cybertivity/omgblockz/dimensions/TheEnd.java
@@ -6,7 +6,7 @@ public class TheEnd extends DimensionBase {
 
     private static final short WORLD_HEIGHT = 256;
     private static final short MAX_LAND_HEIGHT = 256;
-    private static final short SEA_LEVEL = -1;
+    private static final byte SEA_LEVEL = -1;
 
     public TheEnd(int seed, String path) {
         super(Dimension.THE_END, seed,
